+++ src/org/hyperdata/beeps/processors/FIRProcessor.java
@@ -46,7 +46,7 @@ public class FIRProcessor extends DefaultProcessor implements FIRFilter {
 		setFc((Double) getLocal("cutoff"));
 		setnPoints((Integer)  getLocal("npoints"));
 		if (shape.equals("BP") || shape.equals("BS")) {
-			setFc2((Integer)  getLocal("cutoff2"));
+			setFc2((Double)  getLocal("cutoff2"));
 		}
 		initWeights();
 	}
